requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.2
certifi==2023.7.22
charset-normalizer==3.2.0
Django==3.2.18
django-appconf==1.0.5
django-ckeditor==6.7.0
django-crispy-forms==2.0
django-imagekit==4.1.0
django-js-asset==2.1.0
fake-useragent==1.2.1
idna==3.4
importlib-resources==6.0.1
pilkit==2.0
Pillow==10.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
soupsieve==2.4.1
sqlparse==0.4.4
typing-extensions==4.7.1
urllib3==2.0.4
zipp==3.16.2
